Against my gut feeling, moving the part of model that constitutes the language that all services need to understand (protocol), along with topic definitions, to the shared module. This is because both the producer and the consumer of the topic must understand what goes through it.
Perhaps the shared model could be made a "light" version, in case it is possible to simply deserialize a subset of JSON into a simpler structure (e.g. nobody but shipment service cares about the details of Shipment aggregate).
