+++ navigation/src/main/kotlin/com/fedup/navigation/LocationEvent.kt
@@ -1,6 +1,6 @@
 package com.fedup.navigation
 
-import com.fedup.common.*
+import com.fedup.shared.*
 
 sealed class LocationEvent
