+++ navigation/src/main/kotlin/com/fedup/navigation/UserLocationRepository.kt
@@ -1,6 +1,6 @@
 package com.fedup.navigation
 
-import com.fedup.common.*
+import com.fedup.shared.*
 import org.springframework.stereotype.*
 
 /**
