+++ pom.xml
@@ -15,7 +15,7 @@
     </properties>
 
     <modules>
-        <module>common</module>
+        <module>shared</module>
         <module>shipment</module>
         <module>navigation</module>
         <module>user</module>
