+++ shared/src/main/kotlin/com/fedup/shared/CommonTopics.kt
@@ -1,4 +1,4 @@
-package com.fedup.common
+package com.fedup.shared
 
 
 object CommonTopics {
