+++ shared/src/main/kotlin/com/fedup/shared/TrackingId.kt
@@ -1,4 +1,4 @@
-package com.fedup.common
+package com.fedup.shared
 
 import java.util.*
