+++ shared/src/main/kotlin/com/fedup/shared/machinery/KafkaUtils.kt
@@ -1,4 +1,4 @@
-package com.fedup.common.machinery
+package com.fedup.shared.machinery
 
 import org.apache.kafka.clients.consumer.*
 import org.apache.kafka.streams.*
