+++ shipment/src/main/kotlin/com/fedup/shipment/model/User.kt
@@ -1,6 +1,6 @@
 package com.fedup.shipment.model
 
-import com.fedup.common.*
+import com.fedup.shared.*
 
 sealed class User: Entity<String>()
 data class Shipper(override val identity: String): User()
