+++ user/src/main/kotlin/com/fedup/user/PaymentService.kt
@@ -1,7 +1,6 @@
 package com.fedup.user
 
-import com.fedup.common.*
-import org.apache.kafka.common.serialization.*
+import com.fedup.shared.*
 
 class PaymentService {
     /**
