+++ user/src/main/kotlin/com/fedup/user/UserNotificationService.kt
@@ -1,8 +1,5 @@
 package com.fedup.user
 
-import com.fedup.common.*
-import org.apache.kafka.common.serialization.*
-
 /**
  * Subscribes to user-notifications stream and pushes events from this stream to the users they are addressed to
  */
